feat: move `properties.scss` into theme ui
BREAKING CHANGE:

- The DNB properties.scss and properties.js are is moved inside a theme folder /themes/theme-ui/properties.*.
- The CSS package dnb-ui-properties is moved inside a theme folder /themes/theme-ui.
- The following packages do not contain properties anymore. Properties are only a part of a theme file, such as: /style/themes/theme-ui/dnb-theme-ui.*:
  - dnb-ui-basis
  - dnb-ui-body
  - dnb-ui-core
  - dnb-ui-elements
  - dnb-ui-tags
