+++ packages/dnb-design-system-portal/src/docs/uilib/usage/customisation/styling.mdx
@@ -34,7 +34,7 @@ There are several packages you can use and combine.
 - **dnb-ui-tags** - this package will force styles on all the HTML Tags like `<h1>` instead of CSS classes, like `.dnb-h--xx-large`
   - **NB:** needs a `.dnb-core-style` [wrapper class](/uilib/usage/customisation/styling#core-style).
   - Use it carefully - because this will effect existing styles as well.
-- **dnb-ui-properties** - includes only the CSS Custom Properties.
+- **dnb-ui-properties** - includes only the CSS Custom Properties. You find this package is inside `/style/themes/theme-ui/`.
   - There is also a [JavaScript file containing](/uilib/usage/customisation/styling#a-list-of-all-css-properties) the same properties.
 - **dnb-ui-fonts** - includes only the `@font-face` and properties. You find this package is inside `/style/themes/theme-ui/`.
