+++ packages/dnb-eufemia/scripts/prebuild/tasks/makePropertiesFile.js
@@ -56,29 +56,19 @@ export const runFactory = ({ returnResult = false } = {}) =>
     log.start('> PrePublish: transforming style modules')
     try {
       gulp
-        .src(
-          [
-            './src/style/**/properties.scss',
-            '!**/__tests__/**',
-            '!**/stories/**',
-            '!**/*_not_in_use*/**/*',
-          ],
-          {
-            cwd: ROOT_DIR,
-          }
-        )
+        .src(['./src/style/themes/**/properties.scss'], {
+          cwd: ROOT_DIR,
+        })
         .pipe(transform('utf8', transformModulesContent))
         .pipe(
           rename({
-            dirname: './',
-            prefix: '',
             extname: '.js',
           })
-        ) // rename
+        )
         .pipe(
           returnResult
             ? transform('utf8', (result) => resolve(result))
-            : gulp.dest('./src/style', {
+            : gulp.dest('./src/style/themes/', {
                 overwrite: true,
                 cwd: ROOT_DIR,
               })
