+++ packages/dnb-eufemia/src/components/pagination/PaginationBar.tsx
@@ -18,7 +18,7 @@ import PaginationContext from './PaginationContext'
 import Context from '../../shared/Context'
 import Button from '../button/Button'
 import IconPrimary from '../icon-primary/IconPrimary'
-import styleProperties from '../../style/properties'
+import styleProperties from '../../style/themes/theme-ui/properties'
 import { LocaleProps } from '../../shared/types'
 import { SkeletonShow } from '../Skeleton'
