+++ packages/dnb-eufemia/src/style/dnb-ui-body.scss
@@ -4,7 +4,7 @@
  */
 
 @import './core/utilities.scss';
-@import './core/properties.scss';
+
 @import './core/scopes.scss';
 @include globalReset();
