+++ packages/dnb-eufemia/src/style/dnb-ui-core.scss
@@ -5,7 +5,7 @@
 
 // core imports
 @import './core/utilities.scss';
-@import './core/properties.scss';
+
 @import './core/scopes.scss';
 @import './core/helper-classes/helper-classes.scss';
 @include globalReset();
