+++ packages/dnb-eufemia/src/style/dnb-ui-elements.scss
@@ -4,7 +4,6 @@
  */
 
 @import './core/utilities.scss';
-@import './core/properties.scss';
 
 // import all HTML elements
 @import './elements/ui-elements.scss';
