--- packages/dnb-eufemia/src/style/dnb-ui-properties.scss
@@ -1,8 +1,0 @@
-/**
- * Import basis core parts
- *
- */
-
-// core imports
-@import './core/utilities.scss';
-@import './core/properties.scss';
