+++ packages/dnb-eufemia/src/style/dnb-ui-tags.scss
@@ -4,7 +4,6 @@
  */
 
 @import './core/utilities.scss';
-@import './core/properties.scss';
 
 // import and execute all HTML element tags
 @import './elements/ui-tags.scss';
