+++ packages/dnb-eufemia/src/style/themes/theme-ui/dnb-theme-ui.scss
@@ -4,7 +4,7 @@
  */
 
 @import '../../core/utilities.scss';
-@import '../../core/properties.scss';
+@import './properties.scss';
 @import './fonts.scss';
 
 // import all the "class styles" of the elements
