+++ packages/dnb-eufemia/src/style/themes/theme-ui/dnb-ui-properties.scss
@@ -3,5 +3,6 @@
  *
  */
 
+// core imports
 @import '../../core/utilities.scss';
-@import './fonts.scss';
+@import './properties.scss';
