+++ packages/dnb-eufemia/src/style/themes/theme-ui/properties.scss
